// frontend/src/app/core/api/model/budgetDashboardDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BudgetDashboardDto { 
    totalBudget?: number;
    committedFunds?: number;
    actualSpend?: number;
    burndownData?: Array<any>;
}