// frontend/src/app/core/api/model/projectDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProjectDto { 
    id?: number;
    name?: string;
    startDate?: string;
    endDate?: string;
    budget?: number;
    teamName?: string;
}